+++ DatadogSessionReplay/Sources/Feature/SRContextPublisher.swift
@@ -22,7 +22,7 @@ internal class SRContextPublisher {
     }
 
     /// Notifies other Features on the state of Session Replay records count.
-    func setRecordsCountByViewID(_ value: [String: Int64]) {
+    func setRecordsCountByViewID(_ value: [String: Int]) {
         core?.set(baggage: value, forKey: RUMDependency.recordsCountByViewID)
     }
 }
